requirements.txt
openai-agents==0.3.3
deepeval==3.6.3
pytest==8.4.2
python-dotenv == 1.1.1
ruff==0.14.0
black==25.9.0